requirements.txt
# Epitome Production Workbook Generator Dependencies

# Excel generation
xlsxwriter>=3.1.9

# Optional: For enhanced CSV/data processing
pandas>=2.0.0

# Optional: For date handling
python-dateutil>=2.8.2

# LLM API for extraction
google-genai>=0.2.0

# Environment variable management
python-dotenv>=1.0.0

# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# PDF parsing
PyPDF2>=3.0.0
